feat: 删除部分无用函数、注释
