+++ packages/core/src/utils/createApp.ts
@@ -1,18 +1,21 @@
 import { resolveApp } from "@laconic/utils";
 import fs from "fs-extra";
-import { execSync, exec } from "child_process";
-import { confirm, intro, select } from "@clack/prompts";
+import { exec } from "child_process";
+import { confirm } from "@clack/prompts";
 import chalk from "chalk";
-import { join } from "path";
-import ora from "ora";
+import { parse } from "@babel/parser";
+import path from "path";
 
-import { removeDirectory, getNpmPackage } from "./fileController";
-import { ProjectTypes, PackageManagers } from "./questions";
-import { projectLink } from "./constants";
-import isGitInstalled from "./gitCheck";
+import { removeDirectory } from "./fileController";
+import { projectSelect } from "./select";
+import gitCheck from "./gitCheck";
+import Generator from "./Generator";
+import PackageAPI from "./PackageAPI";
+import { createFiles } from "./createFiles";
+import { type Preset, getNpmForPackage } from "./preset";
 import createSuccessInfo from "./createSuccessInfo";
-import createCommitlint from "./createCommitlint";
-import { createPackageJson, createTemplateFile } from "./createFile";
+import dependenciesInstall from "./dependenciesInstall";
+import { createReadmeString } from "./createFile";
 
 // 设置输入模式为原始模式
 process.stdin.setRawMode(true);
@@ -26,81 +29,118 @@ process.stdin.on("data", (key) => {
   }
 });
 
-// 创建项目文件
-const makeDirectory = async (matter, options) => {
-  const rootDirectory = resolveApp(matter);
-  // 如果存在同名文件,且没有输入 -f,
-  if (fs.existsSync(rootDirectory) && !options.force) {
-    const shouldContinue = await confirm({
-      message:
-        "Whether to overwrite a file with the same name that exists in the current directory ?",
-    });
+// 创建项目文件夹
+async function createFolder(rootDirectory: string, options: Record<string, any>) {
+  // 检查目录是否存在
+  if (fs.existsSync(rootDirectory)) {
+    let proceed = options.force; // 如果强制创建，则默认继续
 
-    // 删除已存在文件并创建新文件
-    if (shouldContinue === true) {
-      removeDirectory(matter, true);
-    } else process.exit(1);
+    // 如果不是强制创建，询问用户是否覆盖
+    if (!proceed) {
+      proceed = await confirm({
+        message:
+          "Whether to overwrite a file with the same name that exists in the current directory?",
+      });
+    }
+
+    // 根据用户的选择或强制选项决定是否继续
+    if (proceed) {
+      removeDirectory(rootDirectory, false); // 删除已存在的目录
+    } else {
+      process.exit(1); // 用户选择不覆盖，退出程序
+    }
   }
 
-  execSync(`mkdir ${rootDirectory}`);
-};
-
-// 获取表单结果
-const getTableInfo = async () => {
-  const projectType = (await select({
-    message: "Pick a project type.",
-    options: ProjectTypes,
-  })) as string;
-
-  const packageManageType = (await select({
-    message: "Select the package management tool you will use:",
-    options: PackageManagers,
-  })) as string;
-
-  const commitLint = (await confirm({
-    message: "Pick additional lint features:",
-  })) as boolean;
-
-  return { projectType, packageManageType, commitLint };
-};
+  // 创建目录，如果之前已经删除或目录不存在
+  fs.mkdirSync(rootDirectory, { recursive: true });
+}
 
 // 模板创建主函数
-export default async function createApp(matter: string, options: { force: boolean; dev: boolean }) {
-  intro(chalk.green(" create-you-app "));
-  const rootDirectory = resolveApp(matter);
+export default async function createAppTest(projectName: string, options: Record<string, any>) {
+  // 记录开发环境并设置环境变量
+  process.env.NODE_ENV = options.dev ? "DEV" : "PROD";
 
-  await makeDirectory(matter, options);
+  // 获取到项目的根目录的绝对路径
+  const rootDirectory = resolveApp(projectName);
 
-  const { projectType, packageManageType, commitLint } = await getTableInfo();
+  await createFolder(rootDirectory, options);
 
-  // 依据 projectType 把相关模板 json 写入 package.json 文件
-  fs.writeFileSync(
-    join(rootDirectory, "package.json"),
-    JSON.stringify(createPackageJson(projectType, matter), null, 2),
+  // 获取用户选择预设
+  const preset: Preset = await projectSelect();
+  const { template, packageManager, plugins, buildTool } = preset;
+
+  /* ----------从下面的代码开始，创建package.json---------- */
+  console.log(chalk.blue(`\n📄  Generating package.json...`));
+  // 1. 配置文件基本内容，包含不仅仅是package.json的字段
+  const packageContent = {
+    name: projectName,
+    version: "0.1.0",
+    private: true,
+    devDependencies: {},
+  };
+
+  // 2. 初始化构建工具配置文件
+  const buildToolConfigTemplate = fs.readFileSync(
+    path.resolve(fs.realpathSync(process.cwd()), `./template/${buildTool}.config.js`),
+    "utf-8",
   );
-
-  // 写入 .gitignore 文件
-  fs.writeFileSync(join(rootDirectory, ".gitignore"), createTemplateFile("gitignore"));
-  // 下载 npm 包解压,获取目标模板导入文件,并删除一些无用的代码文件
-  getNpmPackage(projectLink.get(projectType) as string, projectType, rootDirectory, options.dev);
-
-  // 注入 commitlint 规则
-  if (commitLint === true) {
-    createCommitlint(rootDirectory);
-  }
-
-  // todo：考虑省略这一步
-  // 安装相关依赖
-  const spinner = ora().start();
-  spinner.start(chalk.bold.cyan("The dependency package is being installed..."));
-  exec(`${packageManageType} install`, { cwd: rootDirectory }, () => {
-    spinner.succeed(chalk.bold.green("🚀 Project initialization is complete"));
-
-    createSuccessInfo(matter, packageManageType);
+  const buildToolConfigAst = parse(buildToolConfigTemplate, {
+    sourceType: "module",
   });
+  fs.writeFileSync(path.resolve(rootDirectory, `${buildTool}.config.js`), buildToolConfigTemplate);
 
-  // 是否安装已经安装了 git
-  if (isGitInstalled(rootDirectory)) {
-    exec("git init", { cwd: rootDirectory });
+  // 3. 遍历 plugins，插入依赖
+  Object.keys(plugins).forEach((dep) => {
+    console.log("dep:", dep);
+    // TODO: 更多的处理依据 plugins[dep] 后续的变化而插入
+    let { version } = plugins[dep];
+    if (!version) version = "latest"; // 默认版本号为 latest
+    packageContent.devDependencies[dep] = version; // 插件都是以 devDependencies 安装
+    // TODO: 现在只有 babel-plugin-test-ljq 这一个包，先试一下，后续发包
+    if (dep === "Babel") {
+      const pluginName = `${dep.toLowerCase()}-plugin-test-ljq`;
+      packageContent.devDependencies[pluginName] = "latest";
+      delete packageContent.devDependencies["Babel"];
+    }
+  });
+  const packageJson = new PackageAPI(rootDirectory);
+  await packageJson.createPackageJson(packageContent);
+  // 拉取模板
+  // TODO: 新模板未开发，先模拟过程
+  console.log("Creating a project...");
+  //   execSync(`mkdir ${rootDirectory}/src`);
+
+  // 初始化 Git 仓库
+  if (gitCheck(rootDirectory)) exec("git init", { cwd: rootDirectory });
+
+  // 安装传入的依赖
+  if (process.env.NODE_ENV === "PROD") {
+    await dependenciesInstall(rootDirectory, packageManager);
   }
+  // 运行生成器创建项目所需文件和结构
+  console.log(chalk.blue(`🚀  Invoking generators...`));
+  // 传入根目录路径、插件列表、package.json内容创建生成器实例
+  const generators = new Generator(rootDirectory, plugins, packageContent, template, {
+    ast: buildToolConfigAst,
+    buildTool,
+  });
+  await generators.generate();
+
+  // 安装附加依赖
+  // TODO: 待映射部分完成再测试
+
+  await dependenciesInstall(rootDirectory, packageManager);
+  // TODO: configMap 功能目前无用，考虑改为针对于架构的特异化插件选择，目前不影响功能
+  const npmList = getNpmForPackage(preset);
+  console.log("npmList", npmList);
+
+  // 其他剩余操作，如创建 md 文档，或其他首位操作
+  console.log(chalk.blue(`📄  Generating README.md...`));
+  await createFiles(rootDirectory, {
+    "README.md": createReadmeString(packageManager, template, "README.md"),
+    "README-EN.md": createReadmeString(packageManager, template, "README-EN.md"),
+  });
+  createSuccessInfo(projectName, "npm");
+
+  // gitignore
 }
